World wide the number of visitors change since the beginning of the pandemic

The data contains the information of how visitors to (or time spent in) categorized places change compared to the baseline days. A baseline day represents a normal value for that day of the week. The baseline day is the median value from the 5‑week period Jan 3 – Feb 6, 2020.

visitors_total

Format

Data frame with columns

Entity

Name of countries or territories

Code

Country code

Date

Date of the information been recorded

Places

Categorized places

values

Changes compared to our baseline days

Source

Google: Community Mobility Reports https://www.google.com/covid19/mobility/

Examples

library(COVID19dashboard)
visitors_total

#> # A tibble: 173,088 x 5
#>    Entity        Code  Date       Places              values
#>    <chr>         <chr> <date>     <chr>                <dbl>
#>  1 World Average World 2020-02-17 Retail & Recreation  0.962
#>  2 World Average World 2020-02-17 Grocery & Pharmacy   0.934
#>  3 World Average World 2020-02-17 Parks                4.15 
#>  4 World Average World 2020-02-17 Transit Stations     0.587
#>  5 World Average World 2020-02-17 Workplaces           1.64 
#>  6 World Average World 2020-02-17 Residential          0.507
#>  7 World Average World 2020-02-18 Retail & Recreation  0.754
#>  8 World Average World 2020-02-18 Grocery & Pharmacy   0.737
#>  9 World Average World 2020-02-18 Parks                4.21 
#> 10 World Average World 2020-02-18 Transit Stations     0.406
#> # … with 173,078 more rows